Refactor FdTransfer to multiple files & classes, switch to SOCK_SEQPACKET
