+++ src/profiler.cpp
@@ -32,7 +32,7 @@
 #include "itimer.h"
 #include "flameGraph.h"
 #include "flightRecorder.h"
-#include "fdTransfer.h"
+#include "fdTransfer_client.h"
 #include "frameName.h"
 #include "os.h"
 #include "stackFrame.h"
@@ -945,13 +945,8 @@ Error Profiler::start(Arguments& args, bool reset) {
     }
 
     if (args._fdtransfer) {
-        if (!FdTransfer::isListenerInitialized()) {
-            if (!FdTransfer::initializeListener()) {
-                return Error("Failed to initialize FdTransfer listener");
-            }
-        }
-        if (!FdTransfer::acceptPeer()) {
-            return Error("Failed to accept FdTransfer peer");
+        if (!FdTransferClient::connectToServer(OS::processId())) {
+            return Error("Failed to initialize FdTransferClient");
         }
     }
 
@@ -1084,7 +1079,7 @@ Error Profiler::stop() {
     _jfr.stop();
     for (int i = 0; i < CONCURRENCY_LEVEL; i++) _locks[i].unlock();
 
-    FdTransfer::closePeer();
+    FdTransferClient::closePeer();
     _state = IDLE;
     return Error::OK;
 }
